refactor: remove deprecated create commands and implement workload creation
